One writer wizard example creation
