+++ api/internalUtils.go
@@ -40,6 +40,7 @@ type Args struct {
 	ClientId             *string
 	SiteId               *string
 	DeviceId             *string
+	SourceUUID           *string
 	PluginName           bool
 	TimestampGt          *string
 	TimestampLt          *string
@@ -90,6 +91,7 @@ var ArgsType = struct {
 	ClientId             string
 	SiteId               string
 	DeviceId             string
+	SourceUUID           string
 	PluginName           string
 	TimestampGt          string
 	TimestampLt          string
@@ -138,6 +140,7 @@ var ArgsType = struct {
 	ClientId:             "client_id",
 	SiteId:               "site_id",
 	DeviceId:             "device_id",
+	SourceUUID:           "source_uuid",
 	PluginName:           "by_plugin_name",
 	TimestampGt:          "timestamp_gt",
 	TimestampLt:          "timestamp_lt",
